Replace header/body/footer with content
Instead of taking in `Maybe`s for `header`, `body`, and `footer`,
take a single `content: Html msg`. This is nice in a few ways.

First, it Makes Impossible States Impossible, in that we can now never
have an empty modal (`Nothing`s for `header`, `body`, and `footer`)
unless the user specifically sends in an empty `Html msg` as `content`.

Secondly, it mostly decouples the styling from Bootstrap, allowing the
user to enter whatever kind of content they want, with whatever classes
and styles they choose. You can still use Bootstrap, but you no longer
"have" to.